Move robonode_request errors data to shared errors data
