+++ crates/bioauth-flow-rpc/src/errors/shared.rs
